+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 🥷 *Internal*
 * RayRuntime can now be configured to pass resources to underlying remote functions
+* Added version metadata to the workflow IR
 
 
 📃 *Docs*